fixed Select component bubble problem - https://github.com/shadcn-ui/ui/issues/486
